+++ yash-syntax/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to `yash-syntax` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.9.0] - Unreleased
+## [0.9.0] - 2024-06-09
 
 ### Added
 
@@ -243,6 +243,7 @@ command.
 - Functionalities to parse POSIX shell scripts
 - Alias substitution support
 
+[0.9.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.9.0
 [0.8.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.8.0
 [0.7.0]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.7.0
 [0.6.1]: https://github.com/magicant/yash-rs/releases/tag/yash-syntax-0.6.1